% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB.R
\name{calTMB}
\alias{calTMB}
\title{calTMB}
\usage{
calTMB(
  maf,
  bedFile,
  assay = "MSK-v3",
  genelist = NULL,
  mutType = "nonsynonymous",
  bedFilter = TRUE
)
}
\arguments{
\item{maf}{An MAF data frame, generated by \code{\link{vcfToMAF}} function.}

\item{bedFile}{A file in bed format that contains region information.}

\item{assay}{Methodology and assay will be applied as a reference, including
'MSK-v3', 'MSK-v2', 'MSK-v1', 'FoundationOne', 'Pan-Cancer Panel' and
'Customized'. Default: 'MSK-v3'.}

\item{genelist}{A vector of panel gene list, only useful when assay is set to
'Customized'.}

\item{mutType}{A group of variant classifications that will be kept,
only useful when assay is set to 'Pan-Cancer Panel' or 'Customized',
including 'exonic' and 'nonsynonymous'. Default: 'nonsynonymous'.}

\item{bedFilter}{Whether to filter the information in bed file or not, which
only leaves segments in Chr1-Ch22, ChrX and ChrY. Default: TRUE}
}
\value{
A TMB value.
}
\description{
Calculate Tumor Mutational Burden (TMB) in specific regions.
}
\examples{
maf <- vcfToMAF(system.file("extdata", "GC48-2_mutect2.vep.vcf",
package = "CaMutQC"))
TMB_value <- calTMB(maf, bedFile = system.file("extdata/bed",
"Twist_Exome_Target_hg19.bed",package = "CaMutQC"))
}
