% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutFilterCan.R
\name{mutFilterCan}
\alias{mutFilterCan}
\title{mutFilterCan}
\usage{
mutFilterCan(
  maf,
  cancerType,
  tumorDP = 0,
  normalDP = 0,
  tumorAD = 0,
  VAF = 0,
  VAFratio = 0,
  tumorSampleName = "Extracted",
  SBmethod = "SOR",
  SBscore = Inf,
  maxIndelLen = Inf,
  minInterval = 0,
  tagFILTER = NULL,
  dbVAF = 0,
  ExAC = FALSE,
  Genomesprojects1000 = FALSE,
  ESP6500 = FALSE,
  gnomAD = FALSE,
  dbSNP = FALSE,
  COSMIConly = FALSE,
  keepType = "ALL",
  bedFile = NULL,
  bedFilter = TRUE,
  mutFilter = FALSE,
  selectCols = FALSE,
  report = TRUE,
  reportFile = "FilterReport.html",
  reportDir = "./",
  TMB = FALSE
)
}
\arguments{
\item{maf}{An MAF data frame.}

\item{cancerType}{Type of cancer sample whose params needed to be referred to.}

\item{tumorDP}{Threshold of tumor total depth. Default: 20}

\item{normalDP}{Threshold of normal total depth. Default: 10}

\item{tumorAD}{Threshold of tumor alternative allele depth. Default: 10}

\item{VAF}{Threshold of VAF value. Default: 0.05}

\item{VAFratio}{Threshold of VAF ratio (tVAF/nVAF). Default: 5}

\item{tumorSampleName}{Label of the tumor sample, should be one of the
column names of maf. If it is set as 'Extracted', tumorSampleName would be
extracted automatically from the maf data frame. Default: 'Extracted'.}

\item{SBmethod}{Method will be used to detect strand bias,
including 'SOR' and 'Fisher'. Default: 'SOR'. SOR: StrandOddsRatio
(https://gatk.broadinstitute.org/hc/en-us/articles/360041849111-
StrandOddsRatio)}

\item{SBscore}{Cutoff strand bias score used to filter variants.
Default: 3}

\item{maxIndelLen}{Maximum length of indel accepted to be included.
Default: 50}

\item{minInterval}{Maximum length of interval between an SNV and an indel
accepted to be included. Default: 10}

\item{tagFILTER}{Variants with spcific tag in the FILTER column will be kept,
Default: 'PASS'}

\item{dbVAF}{Threshold of VAF of certain population for variants
in database. Default: 0.01}

\item{ExAC}{Whether to filter variants listed in ExAC with VAF higher than
cutoff(set in VAF parameter). Default: TRUE.}

\item{Genomesprojects1000}{Whether to filter variants listed in
Genomesprojects1000 with VAF higher than cutoff(set in VAF parameter).
Default: TRUE.}

\item{ESP6500}{Whether to filter variants listed in ESP6500 with VAF higher
than cutoff(set in VAF parameter). Default: TRUE.}

\item{gnomAD}{Whether to filter variants listed in gnomAD with VAF higher
than cutoff(set in VAF parameter). Default: TRUE.}

\item{dbSNP}{Whether to filter variants listed in dbSNP. Default: FALSE.}

\item{COSMIConly}{Whether to only keep variants in COSMIC. Default: FALSE.}

\item{keepType}{A group of variant classifications will be kept,
including 'exonic' and 'nonsynonymous'. Default: 'exonic'.}

\item{bedFile}{A file in bed format that contains region information.
Default: NULL}

\item{bedFilter}{Whether to filter the information in bed file or not, which
only leaves segments in Chr1-Ch22, ChrX and ChrY. Default: TRUE}

\item{mutFilter}{Whether to directly return a filtered MAF data frame.
If FALSE, a simulation filtration process will be run, and the original MAF
data frame with tags in CaTag column, and  a filter report will be returned.
If TRUE, a filtered MAF data frame and a filter report will be generated.
Default: FALSE}

\item{selectCols}{Columns will be contained in the filtered data frame.
By default (TRUE), the first 13 columns and 'Tumor_Sample_Barcode' column.
Or a vector contains column names will be kept.}

\item{report}{Whether to generate report automatically. Default: TRUE}

\item{reportFile}{File name of the report. Default: 'FilterReport.html'}

\item{reportDir}{Path to the output report file. Default: './'}

\item{TMB}{Whether to calculate TMB. Default: TRUE}
}
\value{
An MAF data frame after common strategy filtration for a cancer type.

A filter report in HTML format
}
\description{
Apply common filter strategies on a MAF data frame for different
cancer types.
}
\examples{
maf <- vcfToMAF(system.file("extdata/Multi-sample",
"SRR3670028.somatic.filter.HC.vep.vcf",package = "CaMutQC"))
mafF <- mutFilterCan(maf, cancerType = 'BRCA')
}
