% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutFilterQual.R
\name{mutFilterQual}
\alias{mutFilterQual}
\title{mutFilterQual}
\usage{
mutFilterQual(
  maf,
  panel = "Customized",
  tumorDP = 0,
  normalDP = 0,
  tumorAD = 0,
  normalAD = Inf,
  VAF = 0,
  VAFratio = 0
)
}
\arguments{
\item{maf}{An MAF data frame, generated by \code{\link{vcfToMAF}} function.}

\item{panel}{The sequencing panel applied on the dataset. Default: "Customized".
Options: "MSKCC", "WES".}

\item{tumorDP}{Threshold of tumor total depth. Default: 0}

\item{normalDP}{Threshold of normal total depth. Default: 0}

\item{tumorAD}{Threshold of tumor alternative allele depth. Default: 0}

\item{normalAD}{Threshold of normal alternative allele depth. Default: Inf}

\item{VAF}{Threshold of VAF value. Default: 0}

\item{VAFratio}{Threshold of VAF ratio (tVAF/nVAF). Default: 0}
}
\value{
An MAF data frame where some variants
has Q tag in CaTag column for sequencing quality filtration
}
\description{
Filter variants in low sequencing quality or low confidence.
}
\examples{
maf <- vcfToMAF(system.file("extdata", "GC48-2_mutect2.vep.vcf",
package = "CaMutQC"))
mafF <- mutFilterQual(maf)
}
