% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutFilterDB.R
\name{mutFilterDB}
\alias{mutFilterDB}
\title{mutFilterDB}
\usage{
mutFilterDB(
  maf,
  VAF = 0.01,
  ExAC = TRUE,
  Genomesprojects1000 = TRUE,
  ESP6500 = TRUE,
  gnomAD = TRUE,
  dbSNP = FALSE,
  COSMIConly = TRUE
)
}
\arguments{
\item{maf}{An MAF data frame, generated by \code{\link{vcfToMAF}} function.}

\item{VAF}{Threshold of VAF value. Default: 0.01}

\item{ExAC}{Whether to filter variants listed in ExAC with VAF higher than
cutoff(set in VAF parameter). Default: TRUE.}

\item{Genomesprojects1000}{Whether to filter variants listed in
Genomesprojects1000 with VAF higher than cutoff(set in VAF parameter).
Default: TRUE.}

\item{ESP6500}{Whether to filter variants listed in ESP6500 with VAF higher
than cutoff(set in VAF parameter). Default: TRUE.}

\item{gnomAD}{Whether to filter variants listed in gnomAD with VAF higher
than cutoff(set in VAF parameter). Default: TRUE.}

\item{dbSNP}{Whether to filter variants listed in dbSNP. Default: FALSE.}

\item{COSMIConly}{Whether to only keep variants in COSMIC. Default: FALSE.}
}
\value{
An MAF data frame after filtration for database
and clinical significance
}
\description{
Filter variants in germline database unless they are flagged
as pathogenic.
}
\examples{
maf <- vcfToMAF(system.file("extdata", "GC48-2_mutect2.vep.vcf",
package = "CaMutQC"))
mafF <- mutFilterDB(maf)
}
