% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutFilterNormalDP.R
\name{mutFilterNormalDP}
\alias{mutFilterNormalDP}
\title{mutFilterNormalDP}
\usage{
mutFilterNormalDP(maf)
}
\arguments{
\item{maf}{An MAF data frame, generated by \code{\link{vcfToMAF}} function.}
}
\value{
An MAF data frame where some variants
has N tag in CaTag column for Normal depth filtration.
}
\description{
Filter dbsnp/non-dbsnp variants based on their normal depth
}
\examples{
maf <- vcfToMAF(system.file("extdata", "GC48-2_mutect2.vep.vcf",
package = "CaMutQC"))
mafF <- mutFilterNormalDP(maf)
}
