% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutFilterPON.R
\name{mutFilterPON}
\alias{mutFilterPON}
\title{mutFilterPON}
\usage{
mutFilterPON(maf)
}
\arguments{
\item{maf}{An MAF data frame, generated by \code{\link{vcfToMAF}} function.}
}
\value{
An MAF data frame after PON filtration.
}
\description{
Filter variants in Panel of Normals
}
\examples{
maf <- vcfToMAF(system.file("extdata", "GC48-2_mutect2.vep.vcf",
package = "CaMutQC"))
mafF <- mutFilterPON(maf)
}
