% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutFilterQual.R
\name{mutFilterQual}
\alias{mutFilterQual}
\title{mutFilterQual}
\usage{
mutFilterQual(
  maf,
  tumorDP = 20,
  normalDP = 10,
  tumorAD = 10,
  VAF = 0.05,
  VAFratio = 5
)
}
\arguments{
\item{maf}{An MAF data frame, generated by \code{\link{vcfToMAF}} function.}

\item{tumorDP}{Threshold of tumor total depth. Default: 20}

\item{normalDP}{Threshold of normal total depth. Default: 10}

\item{tumorAD}{Threshold of tumor alternative allele depth. Default: 10}

\item{VAF}{Threshold of VAF value. Default: 0.05}

\item{VAFratio}{Threshold of VAF ratio (tVAF/nVAF). Default: 5}
}
\value{
An MAF data frame where some variants
has Q tag in CaTag column for sequencing quality filtration
}
\description{
Filter variants in low sequencing quality or low confidence.
}
\examples{
maf <- vcfToMAF(system.file("extdata", "GC48-2_mutect2.vep.vcf",
package = "CaMutQC"))
mafF <- mutFilterQUal(maf)
}
