% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutFilterReg.R
\name{mutFilterReg}
\alias{mutFilterReg}
\title{mutFilterReg}
\usage{
mutFilterReg(maf, bedFile = NULL, bedHeader = FALSE, bedFilter = TRUE)
}
\arguments{
\item{maf}{An MAF data frame, generated by \code{\link{vcfToMAF}} function.}

\item{bedFile}{A bed file that contains region information.
Default: NULL}

\item{bedHeader}{Whether the input bed file has a header or not. 
Default: FALSE.}

\item{bedFilter}{Whether to filter the information in bed file or not, which
only leaves segments in Chr1-Ch22, ChrX and ChrY. Default: TRUE}
}
\value{
An MAF data frame where some variants
have R tag in CaTag column for region filtration.
}
\description{
Filter variants in specific regions.
}
\examples{
maf <- vcfToMAF(system.file("extdata", "GC48-2_mutect2.vep.vcf",
package = "CaMutQC"))
mafF <- mutFilterReg(maf, bedFile = system.file("extdata/bed/Twist",
"Twist_Exome_Target_hg19.bed",package = "CaMutQC"))
}
